unit test structured.go 4: existing test parity
